+++ cheetah/accelerator/vertical_corrector.py
@@ -40,11 +40,14 @@ class VerticalCorrector(Element):
         factory_kwargs = {"device": device, "dtype": dtype}
         super().__init__(name=name)
 
-        self.length = torch.as_tensor(length, **factory_kwargs)
-        self.angle = (
-            torch.as_tensor(angle, **factory_kwargs)
-            if angle is not None
-            else torch.zeros_like(self.length)
+        self.register_buffer("length", torch.as_tensor(length, **factory_kwargs))
+        self.register_buffer(
+            "angle",
+            (
+                torch.as_tensor(angle, **factory_kwargs)
+                if angle is not None
+                else torch.zeros_like(self.length)
+            ),
         )
 
     def transfer_map(self, energy: torch.Tensor) -> torch.Tensor: